DRILL-5721: Refactor for FragmentManager, setupRootFragment and startNewFragment
DRILL-5721: Query with only root fragment and no non-root fragment hangs when Drillbit to Drillbit Control Connection has network issues Note: To resolve the issue all the fragments including root fragment which are assigned to be executed on Foreman node are scheduled locally and not sent over Control Tunnel. Also the FragmentStatusReporter is updated to sent the status update locally by fragments running on Foreman node.

DRILL-5721: Update the test added for DRILL-5701 as there is change in behavior

close apache/drill#919